Recreate /var/run/influxdb dir upon recovery
This update fixes an issue where the /var/run/influxdb directory
is not being re-created over a DOR because the controller manifest that
creates it is not being run in that recovery mode.

The fix is to enhance the influxdb service file to ensure this directory
is created whenever the service is started.

Story: 2002823
Task: 22740

Change-Id: Iecd81969ae1611b963fae5595f60c3eb2d2da851
Signed-off-by: Jack Ding <jack.ding@windriver.com>