DPDK: mlx5 driver: fix memory region cache lookup
This commit backports commit 95086c6 from upstream DPDK, as well
as adds a similar check for a valid memory region.

https://github.com/DPDK/dpdk/commit/95086c6

Since the latest mellanox driver has had a lot of refactoring in
the latest upstream, this patch, as well as the associated bug
should be re-evaluated on the next openvswitch/DPDK upversion.

The original commit states:

The Memory Region (MR) cache contains pointers to mlx5_mr.
The MR cache indexes are filled when a new MR is created. As it is
possible for MR to be created on the flight, an extra validation must be
added to avoid segmentation fault.

Change-Id: Ia54267a7fab30dbde96f4601ca33f0ca117a0650
Closes-Bug: #1792418
Signed-off-by: Steven Webster <steven.webster@windriver.com>