Fix exec check which restricts ceph-prepare execution
Minor patch update to the command that checks to see if an OSD should be
erased.

The recent bug fix attempts to perform command expansion within a
literal string. This will always fail and cause the OSD to be
reformatted.

Change the single quotes to double quotes to allow the command expansion
to execute successfully.

Change-Id: I522207e7e5f7a428fb3a962ab6a13574f73bb3c9
Related-Bug: #1800889
Signed-off-by: Robert Church <robert.church@windriver.com>