refactor syslog-ng
Package syslog-ng-config is created to install customized
config file of syslog-ng. Since there is no source code change
in syslog-ng, we could replace the srpm with rpm directly.

syslog-ng-config is set to depends on syslog-ng. So syslog-ng
will be installed automatically.

Test:
Pass build and basic deploy test. Confirmed the related config
file is the same as before.

Story: 2003768
Task: 27599
Depends-On: https://review.openstack.org/616720

Change-Id: I2a4e15b9ffde92aa59072d590de2b56d239e29ad
Signed-off-by: slin14 <shuicheng.lin@intel.com>