IPaddr2 resource agent fix for loopback interface
The IPaddr2 resource agent currently assigns addresses
to the loopback interface without specifying the scope,
so they default to the global scope. This fails if the
address is not in the loopback address range.

The fix is to always specify "scope host" when assigning
addresses to the loopback interface.

Change-Id: Ib7c08bfc1c661a9fb347308e55cc806b4dd129ad
Story: 2003910
Task: 26796
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>