Fix remote logging traffic control filter priority
Previous commit 01f5fdd made a required change to filter
infrastructure traffic on the management interface with an 802.1q
protocol in the case of a consolidated interface.

However, this has caused the remote logging tc script to have a
failure.  The script tries to install 'ip' protocol filters at the
same priority as the 802.1q filters, which is rejected by the
kernel.

This commit detects a consolidated interface situation and bumps
the priority of the remote logging tc filter priority on the
management interface, similarly to what is done in the main
cgcs_tc_setup script.

The file has also been cleaned up to pass bashate.

Related-Bug: #1807055
Change-Id: Id11625c0f9bcbf109f574563ff284d4a36bc6377
Signed-off-by: Steven Webster <steven.webster@windriver.com>