build mariadb docker image with galera arbitrator added
In order to support running the galera arbitrator we need to add
it to the openstack-helm mariadb docker image.  This means building
our own docker image for now.

I've talked with "jayahn" on the openstack-helm IRC channel, and they
said they had no objection to adding galera-arbitrator to their mariadb
image, so we should upstream it as soon as possible.  Once it's
upstreamed we can remove this.

Change-Id: I6ab2607abcd8e0d130ef80fbd1979c62a20a6ff4
Story: 2004712
Task:  29053
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>