Ceph build script improvements to prevent needless rebuilds
Problem: The file $SRC_DIR/src/.git_version is created by the
custom build script and is not cleaned upon exit.  The presence
of the new file under $SRC_DIR will trigger a rebuild on the
next iteration.

Solution: Add a cleanup of the file to cover the normal exit case.
This means $SRC_DIR/src/.git_version will not be present at the
start of the next build.

Problem: A lot of tarballs are copied into the build by the
build script, rather than being listed in the COPY_LIST.  This
breaks the md5 checksum mechanism used to determine if a rebuild
is required.  A change to a tarball will be ignored and not
trigger a rebuild.

Solution: Move the code that generates the list of input tarballs
into build_srpm.data.  The file is sourced, so COPY_LIST can
be populated dynamically.

Problem: Script returns success when rpmbuild fails.

Solution: Propogate the error code to the final exit.

Change-Id: I2e760c24ecd3ce2d237863b948863c2a876d24fa
Closes-Bug: 1830130
Co-authored-by: Shuicheng Lin <shuicheng.lin@intel.com>
Signed-off-by: Scott Little <scott.little@windriver.com>