speclint: Add script for linting RPM specfiles
This script along with the macros file will be used to initially
check the specfiles being added as part of the MultiOS work. It
will be executed via zuul and tox to scan the openSUSE specfiles,
later we can start add the CentOS specfiles also.

The macros.openstack-singlespec is orignally from
openstack/rpm-packaging customized with tis_patch_ver

Change-Id: I22bf778388e6dcc3ca42d5fd6ec16b30200d6b75
Signed-off-by: Saul Wold <sgw@linux.intel.com>