Disable CONFIG_MEMCG_KMEM to avoid slab leak
The kernel memory accounting in the RHEL kernel is broken and
results in a slab memory leak when it is enabled. See the
following bug for details:
https://bugzilla.redhat.com/show_bug.cgi?id=1507149

Unfortunately, this option is enabled by default, so it must be
disabled. Even worse, the kernel won't compile with the option
disabled, so a fix for the compile error is also included.

Change-Id: I627106ae25f86204c1954c1c21171bbef348afaf
Closes-Bug: 1835534
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>